Docs/show docsidebar on mobile (#878)
