+++ docs/public/index.css
@@ -115,12 +115,25 @@ article > section :is(ul, ol) > * + * {
   margin-top: 0.75rem;
 }
 
+article > section nav :is(ul, ol) > * + * {
+  margin-top: inherit;
+}
+
 article > section li > :is(p, pre, blockquote):not(:first-child) {
   margin-top: 1rem;
 }
 
 article > section :is(ul, ol) {
   padding-left: 1em;
+}
+
+article > section nav :is(ul, ol) {
+  padding-left: inherit;
+}
+
+article > section nav {
+  margin-top: 1rem;
+  margin-bottom: 2rem;
 }
 
 article > section ::marker {
